+++ packages/lavas-core-vue/core/middlewares/express-error.js
@@ -6,35 +6,28 @@
 /**
  * generate error middleware
  *
- * @param {string} errPath errPath
+ * @param {string} option.errPath errPath
  * @return {Function} koa middleware
  */
-export default function (errPath) {
+export default function ({errorPath, defaultErrorMessage, showRealErrorMessage}) {
 
     return async (err, req, res, next) => {
         if (req.lavasIgnoreFlag) {
             return next();
         }
 
-        let errorMsg = 'Internal Server Error';
-        if (err.status !== 404) {
-            console.log('[Lavas] error middleware catch error:');
-            console.log(err);
-        }
-        else {
-            errorMsg = `${req.url} not found`;
-            console.log(errorMsg);
-        }
+        console.log('[Lavas] error middleware catch error:');
+        console.log(err);
 
-        if (errPath === req.url.replace(/\?.+$/, '')) {
+        if (errorPath === req.url.replace(/\?.+$/, '')) {
             // if already in error procedure, then end this request immediately, avoid infinite loop
             res.end();
             return;
         }
 
         // redirect to the corresponding url
-        let target = `${errPath}?error=${encodeURIComponent(errorMsg)}`;
-        if (errPath) {
+        let target = `${errorPath}?error=${encodeURIComponent(showRealErrorMessage ? err.message : defaultErrorMessage)}`;
+        if (errorPath) {
             res.writeHead(302, {Location: target});
         }