+++ packages/lavas-core-vue/core/middlewares/ssr.js
@@ -3,9 +3,7 @@
  * @author lavas
  */
 
-import {readFile} from 'fs-extra';
-import {join} from 'path';
-import {matchUrl} from '../utils/router';
+import Logger from '../utils/logger';
 
 /**
  * generate ssr middleware
@@ -14,7 +12,6 @@ import {matchUrl} from '../utils/router';
  * @return {Function} koa middleware
  */
 export default function (core) {
-    let {cwd, config, renderer, builder, isProd} = core;
     return async function (req, res, next) {
         if (req.lavasIgnoreFlag) {
             return next();
@@ -23,9 +20,9 @@ export default function (core) {
         let url = req.url;
         let errorHandler = err => next(err);
 
-        console.log(`[Lavas] route middleware: ssr ${url}`);
+        Logger.info('route middleware', `ssr ${url}`);
 
-        let {err, html} = await renderer.render({
+        let {err, html} = await core.renderer.render({
             url,
             req,
             res,